+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Remove lang prefix from link-autocomplete hrefs to prevent double prefixes
+
 ## [0.2.0-alpha.11] - 2026-05-18
 
 ### Added